refactor(leo,wasm): share EnvOptions/BuildOptions via leo-lang::options
`crates/leo`'s lib is now wasm-buildable. The native CLI surface (cli/,
errors/) stays gated behind `#[cfg(not(target_arch = "wasm32"))]` — that
tree still pulls in snarkVM's umbrella, tokio, axum, reqwest, terminal
UI, and the rest of the native toolbelt. What survives the cfg gate is
a new top-level `pub mod options` carrying the wasm-shared option
structs.

`EnvOptions` and `BuildOptions` move from
`cli/commands/common/options.rs` into `crate::options`; all their
fields become `pub` (were `pub(crate)`). The CLI module re-exports
them via `pub use crate::options::*` so every existing
`cli::commands::*` import keeps working unchanged.

`leo-wasm` now depends on `leo-lang` and re-exports `EnvOptions` /
`BuildOptions` from `leo_lang::options::*` via `crate::wire`. The
parallel `EnvOptions` definition that lived in `leo-wasm/src/wire.rs`
is gone. JSON `from_json` and `resolved_network()` helpers live on the
shared struct so both surfaces consume it the same way.

Cargo.toml: native-only deps in `crates/leo` are now under
`[target.'cfg(not(target_arch = "wasm32"))'.dependencies]`. `getrandom`
gets a `wasm32`-only entry with the `wasm_js` backend (matches the
existing pattern in `leo-package`).

Follow-up: lift `handle_build`/`handle_run`/`handle_test` into the
shared `options` module's neighbour so the leo-wasm command modules
become thin wasm-bindgen shims over the CLI's command impls.
