+++ Cargo.toml
@@ -11,6 +11,7 @@ leo-compiler        = { path = "./crates/compiler",         version = "=4.1.0" }
 leo-disassembler    = { path = "./crates/disassembler",     version = "=4.1.0" }
 leo-errors          = { path = "./crates/errors",           version = "=4.1.0" }
 leo-fmt             = { path = "./crates/fmt",              version = "=4.1.0" }
+leo-lang            = { path = "./crates/leo",              version = "=4.1.0" }
 leo-lsp             = { path = "./crates/lsp",              version = "=4.1.0" }
 leo-package         = { path = "./crates/package",          version = "=4.1.0" }
 leo-parser          = { path = "./crates/parser",           version = "=4.1.0" }
